requirements.txt
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
exceptiongroup==1.2.0
google-api-core==2.17.1
google-api-python-client==2.122.0
google-auth==2.28.2
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
h11==0.14.0
httplib2==0.22.0
idna==3.6
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pkg_resources==0.0.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.2
PySocks==1.7.1
pyTelegramBotAPI==4.16.1
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
selenium==4.18.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
webdriver-manager==4.0.1
wsproto==1.2.0
